Made search function work and error handling

+++ app/src/main/java/Models/Data.java
@@ -1,7 +1,5 @@
 package Models;
 
-import android.os.StrictMode;
-
 import com.esri.arcgisruntime.geometry.Geometry;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
@@ -22,10 +20,10 @@ public class Data {
     private BigInteger object_id;
     @SerializedName("long")
     @Expose
-    private BigDecimal longitude;
+    private float longitude;
     @SerializedName("lat")
     @Expose
-    private BigDecimal latitude;
+    private float latitude;
     @SerializedName("x_utm")
     @Expose
     private BigInteger x_utm;
@@ -33,7 +31,7 @@ public class Data {
     @Expose
     private BigInteger y_utm;
 
-    public Data(Integer gid, BigInteger object_id, BigDecimal longitude, BigDecimal latitude,
+    public Data(Integer gid, BigInteger object_id, float longitude, float latitude,
                 BigInteger x_utm, BigInteger y_utm, String x_y_utm, String street_nam, String house_num,
                 String kommune, BigDecimal no2_street, BigInteger idnr, Geometry geom) {
         this.gid = gid;
@@ -89,19 +87,19 @@ public class Data {
         this.object_id = object_id;
     }
 
-    public BigDecimal getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(BigDecimal longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
 
-    public BigDecimal getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(BigDecimal latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
